test: move NavigatorContext and update test factory imports
