+++ src/screens/Gurbani/Lines.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { fireEvent, render } from '@testing-library/react-native'
 
-import factories from '../../../test/factories'
+import * as factories from '../../../test/factories'
 
 import Lines from './Lines'
