+++ src/screens/Search/Navbar.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { fireEvent, render } from '@testing-library/react-native'
 
-import wrapper from '../../lib/NavigatorContext'
+import wrapper from '../../../test/utils/NavigatorContext'
 
 import SearchNavbar from './Navbar'
