+++ test/factories/index.ts
@@ -2,4 +2,3 @@ export * from './line'
 export * from './shabad'
 export * from './search'
 export * from './bookmark'
-
